implement drawer content with starter copy
