+++ packages/client/modules/userDashboard/components/OrgBilling/OrgPlanDrawer.tsx
@@ -13,8 +13,7 @@ import useBreakpoint from '../../../../hooks/useBreakpoint'
 import PlainButton from '../../../../components/PlainButton/PlainButton'
 import useAtmosphere from '../../../../hooks/useAtmosphere'
 import {ICON_SIZE} from '../../../../styles/typographyV2'
-import {EnterpriseBenefits, TeamBenefits} from '../../../../utils/constants'
-import {upperFirst} from '../../../../utils/upperFirst'
+import OrgPlanDrawerContent from './OrgPlanDrawerContent'
 
 const DrawerHeader = styled('div')({
   alignItems: 'center',
@@ -55,24 +54,6 @@ const CloseIcon = styled(Close)({
   }
 })
 
-const List = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-  overflowY: 'auto',
-  padding: '16px 0px',
-  position: 'relative',
-  width: '100%'
-})
-
-const DrawerContent = styled('div')({
-  backgroundColor: PALETTE.WHITE,
-  display: 'flex',
-  overflow: 'hidden',
-  padding: 16,
-  height: '100vh',
-  flexDirection: 'column'
-})
-
 const StyledCloseButton = styled(PlainButton)({
   height: ICON_SIZE.MD18,
   display: 'flex',
@@ -84,73 +65,6 @@ const StyledLabelHeading = styled(LabelHeading)({
   lineHeight: '18px',
   textTransform: 'none'
 })
-
-const Title = styled('span')({
-  fontWeight: 600,
-  fontSize: 20
-})
-
-const Subtitle = styled('span')({
-  fontWeight: 600,
-  paddingBottom: 8
-})
-
-const UL = styled('ul')({
-  margin: 0
-})
-
-const LI = styled('li')<{isBlue?: boolean}>(({isBlue}) => ({
-  fontSize: 16,
-  lineHeight: '28px',
-  color: isBlue ? PALETTE.SKY_500 : PALETTE.SLATE_900,
-  textTransform: 'none',
-  fontWeight: 400,
-  textAlign: 'left',
-  listStyleType: 'disc'
-}))
-
-const Link = styled('a')({
-  color: PALETTE.SKY_500,
-  fontWeight: 600,
-  textDecoration: 'none',
-  '&:hover': {
-    color: PALETTE.SKY_500,
-    textDecoration: 'underline'
-  }
-})
-
-const agileResources = [
-  {
-    title: '57 Daily Standup Questions for More Engaging Updates',
-    url: 'https://www.parabol.co/resources/daily-standup-questions/'
-  },
-  {
-    title: '29 Effective Meeting Tips for Advanced Facilitators',
-    url: 'https://www.parabol.co/blog/effective-meeting-tips/'
-  },
-  {
-    title: '50+ Retrospective Questions for your Next Meeting',
-    url: 'https://www.parabol.co/resources/retrospective-questions/'
-  },
-  {
-    title: '8 Agile Estimation Techniques to Try With your Team',
-    url: 'https://www.parabol.co/blog/agile-estimation-techniques/'
-  }
-]
-
-const enterpriseResources = {
-  retros: 'https://www.parabol.co/agile/retrospectives/',
-  estimation: 'https://www.parabol.co/agile/sprint-poker/',
-  standups: 'https://www.parabol.co/agile/online-standup-tool/',
-  templates: 'https://www.parabol.co/templates/',
-  integrations: 'https://www.parabol.co/integrations/'
-}
-
-const featuresLookup = {
-  starter: [],
-  team: TeamBenefits,
-  enterprise: EnterpriseBenefits
-} as const
 
 type Props = {
   organizationRef: OrgPlanDrawer_organization$key
@@ -180,6 +94,7 @@ const OrgPlanDrawer = (props: Props) => {
       org.setValue(!showDrawer, 'showDrawer')
     })
   }
+
   return (
     <ResponsiveDashSidebar
       isOpen={showDrawer}
@@ -194,77 +109,7 @@ const OrgPlanDrawer = (props: Props) => {
             <CloseIcon />
           </StyledCloseButton>
         </DrawerHeader>
-        <DrawerContent>
-          <Title>{`🎉 Welcome to the ${upperFirst(tier)} Plan!`}</Title>
-          <List>
-            <Subtitle>
-              {'In addition to the Parabol features you’re used to, you now have access to: '}
-            </Subtitle>
-            <UL>
-              {featuresLookup[tier].map((feature) => (
-                <LI key={feature}>{feature}</LI>
-              ))}
-            </UL>
-          </List>
-          <List>
-            <Subtitle>{'Resources for effective agile teams:'}</Subtitle>
-            <UL>
-              {agileResources.map((resource) => (
-                <LI isBlue key={resource.title}>
-                  <Link href={resource.url} target='_blank' rel='noopener noreferrer'>
-                    {resource.title}
-                  </Link>
-                </LI>
-              ))}
-            </UL>
-          </List>
-          {tier === 'enterprise' && (
-            <List>
-              <Subtitle>{'Make the most out of Parabol:'}</Subtitle>
-              <UL>
-                <LI>
-                  {'Cover all of your agile meetings - '}
-                  <Link href={enterpriseResources.retros} target='_blank' rel='noopener noreferrer'>
-                    {'retros, '}
-                  </Link>
-                  <Link
-                    target='_blank'
-                    rel='noopener noreferrer'
-                    href={enterpriseResources.estimation}
-                  >
-                    {'estimation'}
-                  </Link>
-                  {' & '}
-                  <Link
-                    target='_blank'
-                    rel='noopener noreferrer'
-                    href={enterpriseResources.standups}
-                  >
-                    {'standups'}
-                  </Link>
-                </LI>
-                <LI>
-                  <Link
-                    target='_blank'
-                    rel='noopener noreferrer'
-                    href={enterpriseResources.templates}
-                  >
-                    {'40+ Meeting Templates'}
-                  </Link>
-                </LI>
-                <LI>
-                  <Link
-                    target='_blank'
-                    rel='noopener noreferrer'
-                    href={enterpriseResources.integrations}
-                  >
-                    {'Integrate with Jira, Slack & More'}
-                  </Link>
-                </LI>
-              </UL>
-            </List>
-          )}
-        </DrawerContent>
+        <OrgPlanDrawerContent tier={tier} />
       </Drawer>
     </ResponsiveDashSidebar>
   )
